offline shader strategy

+++ components/shaders/include/shaders/reflectors/spirv_reflector.hpp
@@ -12,7 +12,7 @@ class SpirvReflector
 	SpirvReflector()  = default;
 	~SpirvReflector() = default;
 
-	ShaderResourceSet reflect(const std::vector<uint32_t> &code) const;
+	ShaderResourceSet reflect(const std::vector<uint8_t> &code) const;
 
   private:
 	std::vector<ShaderResource> reflect_input_variables(const SpvReflectShaderModule &module) const;
